requirements.txt
py-clob-client>=0.17.0
anthropic>=0.49.0
requests>=2.31.0
python-dotenv>=1.0.0
eth-account>=0.13.0
web3>=7.0.0
